+++ src/open_refinery/__init__.py
@@ -150,7 +150,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 __all__ = [
     "Factory",